refactors postgres functions into new PostgreSQLSource class, cleans up setup.py and requirments (remove docx and biopython), updates to README
