+++ README.md
@@ -29,20 +29,29 @@ like [Protege](http://protege.stanford.edu/).
 * One of the unit tests requires
 [owltools](https://code.google.com/p/owltools/wiki/InstallOWLTools) be available on your path.  You could modify
 the code to skip this, if necessary
-* unit tests require nosetests (if on OS X you may need to `sudo pip3 install nose`)
+* Running make test requires nosetests (if on OS X you may need to `sudo pip3 install nose`)
 
 * Required external python packages:
-    * [psycopg2](http://initd.org/psycopg/) 
     * [rdflib](https://code.google.com/p/rdflib/)
     * isodate
     * roman
-    * [python-docx](https://github.com/python-openxml/python-docx)
     * pyyaml
-    * pysftp
-    * [biopython](https://github.com/biopython/biopython)
-    * docx
+
+    
+* Optional source specific python packages:
+    * [psycopg2](http://initd.org/psycopg/)
+    * [python-docx](https://github.com/python-openxml/python-docx)
     * beautifulsoup4
     * GitPython
+    * intermine
+    * pysftp
+    
+Note, Dipper imports source modules dynamically at runtime.  As a result it is possible to build a core set
+of requirements and add source specific dependencies as needed.  Presently this only implemented with pip requirements
+files. For example to build dependencies for MGI:
+
+        pip3 install -r requirements/core.txt
+        pip3 install -r requirements/mgi.txt
     
 If you encounter any errors installing these packages using Homebrew, it could be due to [a curent known issue in upgrading to  pip3](https://github.com/Homebrew/homebrew/issues/25752). In this case, first force reinstall pip2 (````pip2 install --upgrade --force-reinstall pip````) and then install the package using pip3 (eg. ````pip3 install psycopg2````.)
