+++ requirements/core.txt
@@ -1,0 +1,6 @@
+rdflib
+isodate
+roman
+python-docx
+pyyaml
+biopython
