+++ requirements/coriell.txt
@@ -1,0 +1,1 @@
+pysftp
