+++ requirements/genereviews.txt
@@ -1,0 +1,1 @@
+beautifulsoup4
