+++ requirements/mgi.txt
@@ -1,0 +1,1 @@
+psycopg2
